requirements.txt
boto3==1.28.11
botocore==1.31.11
gitdb==4.0.10
GitPython==3.1.32
jmespath==1.0.1
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
smmap==5.0.0
urllib3==1.26.16
